Read me
The attached dataset and code files are used for the reported results in the paper entitled "Predicting Stock Prices Using Ensemble Learning Techniques". The code is written by Salama Elsayed under the supervision of Prof. Ibrahim Elhnawy, Prof. Ahmad Salah, and Dr. Marwa Abdella. The paper is part of the requirement for Salama to fulfill her MSc of CS from Zagazig University.

The paper is under review in the International Journal of Electrical and Computer Engineering (IJECE).